backend/requirements/dev.txt
-r common.txt
astroid==1.5.3
factory_boy
isort
pylint==1.7.5
pylint-django
pytest-cov
pytest-django
snapshottest
